third_party/blink/web_tests/wpt_internal/css/css-size-adjust/os-font-scale-required/text-size-adjust-env-affects-font-size-dynamic.html

<!DOCTYPE html>
<meta content="width=device-width,initial-scale=1.0,minimum-scale=1.0"
  name="viewport">
<meta name="assert"
  content="text-size-adjust should affect font-size when used with env(preferred-text-scale), even for large values.">
<link rel="help"
  href="https://drafts.csswg.org/css-size-adjust/#propdef-text-size-adjust">
<link rel="help"
  href="https://github.com/w3c/csswg-drafts/issues/10674#issuecomment-2625414100">
<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>
<script>
  function runTest() {
    test(() => {
      const fontSize = getComputedStyle(textAdjust2FontSize12)['font-size'];
      assert_equals(fontSize, '24px');
    }, 'Font-size on element with text-size-adjust: 200% and font-size 12px');

    test(() => {
      const fontSize = getComputedStyle(textAdjust2FontSize16)['font-size'];
      assert_equals(fontSize, '32px');
    }, 'Font-size on element with text-size-adjust: 200% and font-size 16px');

    test(() => {
      const fontSize = getComputedStyle(textAdjust2FontSize20)['font-size'];
      assert_equals(fontSize, '40px');
    }, 'Font-size on element with text-size-adjust: 200% and font-size 20px');

    if (window.internals) {
      internals.settings.setAccessibilityFontScaleFactor(1.0);
      document.body.left;
      test(() => {
        const fontSize = getComputedStyle(textAdjust2FontSize12)['font-size'];
        assert_equals(fontSize, '12px');
      }, 'Font-size on element with text-size-adjust: 100% and font-size 12px');

      test(() => {
        const fontSize = getComputedStyle(textAdjust2FontSize16)['font-size'];
        assert_equals(fontSize, '16px');
      }, 'Font-size on element with text-size-adjust: 100% and font-size 16px');

      test(() => {
        const fontSize = getComputedStyle(textAdjust2FontSize20)['font-size'];
        assert_equals(fontSize, '20px');
      }, 'Font-size on element with text-size-adjust: 100% and font-size 20px');
    } else {
      document.body.appendChild(document.createTextNode("No window.internals? This test probably failed already anyway."));
    }
  }
</script>

<body onload="runTest()"
  style="text-size-adjust: calc(100% * env(preferred-text-scale));">
  <div id="textAdjust2FontSize12" style="font-size: 12px;">Hello world</div>
  <div id="textAdjust2FontSize16" style="font-size: 16px;">Hello world</div>
  <div id="textAdjust2FontSize20" style="font-size: 20px;">Hello world</div>
</body>
